requirements.txt
flask
paho-mqtt
httpx